requirements.txt
python-telegram-bot==20.7
matplotlib==3.8.2
numpy==1.26.2
python-dotenv==1.0.0
